refactor: 유틸파일 lib폴더로 분해, test폴더 삭제
