+++ src/components/galleryExhibition/threejs/Ceiling.tsx
