+++ src/components/galleryExhibition/threejs/Floor.tsx
