+++ src/components/galleryExhibition/threejs/Player.tsx
@@ -1,7 +1,7 @@
 import { useFrame } from '@react-three/fiber';
 import { useEffect, useMemo, useRef } from 'react';
 import * as THREE from 'three';
-import { WAllType } from '../../../../types/gallery';
+import { WAllType } from '../../../types/gallery';
 
 export default function Player({
   size = 15,
